% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Exp}
\alias{Exp}
\title{Exponential Transformation}
\usage{
Exp(x, k)
}
\arguments{
\item{x}{integer, value to be exponentiated}

\item{k}{integer, number of terms to stop at}
}
\value{
Returns exponential of x to sum of k terms
}
\description{
Gives the exponential of a number, truncated at a sum of k terms.
}
\details{
This function will take two integer values, x and k, and return the exponential of x to a sum of k terms
}
\examples{
Exp(5, 5)
Exp(10, 15)

}
