% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{make_p3_class}
\alias{make_p3_class}
\title{P3 Class Constructor Function}
\usage{
make_p3_class(x)
}
\arguments{
\item{x, }{A list or vector.}
}
\value{
Returns nothing
}
\description{
Constructs a P3 class
}
\details{
This function creates a class, p3_class, from an object passed into the function, and turns it into a list
}
\examples{
h <- rnorm(100)
p3 <- make_p3_class(h)
class(p3)
}
