% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{print.p3_class}
\alias{print.p3_class}
\title{P3 Class Print Method}
\usage{
\method{print}{p3_class}(x, ...)
}
\arguments{
\item{x, }{p3_class}

\item{..., }{Can use to pass arguments to the print function}
}
\value{
Returns a phrase containing the class and number of observations
}
\description{
Creates a print method for the p3_class
}
\details{
This function will create a print method for the p3_class, and then return the class name and number of observations.
}
\examples{
h <- rnorm(100)
p3 <- make_p3_class(h)
class(p3)
print(p3)
}
