% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{calculate_CI}
\alias{calculate_CI}
\title{Calculate Confidence Interval}
\usage{
calculate_CI(x, conf)
}
\arguments{
\item{x, }{vector of length X}

\item{conf, }{number between 0-1 that determines confidence interval \%.}
}
\value{
Returns vector of length 2, with the first term as the lower bound and 2nd term as the upper bound of the CI
}
\description{
Calculate the confidence interval of the mean for a vector
}
\details{
This function will take a vector of length X and calculate the confidence interval of the mean of that vector. The alpha value is adjusted with the second term, which ranges from 0-1
}
\examples{
X <- c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10)
calculate_CI(X, 0.95) # X is a vector of length N, 0.95 is for a 95\% CI
calculate_CI(X, 0.9) # X is a vector of length N, 0.9 is for a 90\% CI

}
