% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3Class.R
\name{print.p3_class}
\alias{print.p3_class}
\title{P3 Class Print Method}
\usage{
\method{print}{p3_class}(x)
}
\arguments{
\item{x}{integer, value to be exponentiated}

\item{k}{integer, number of terms to stop at}
}
\value{
Returns exponential of x to sum of k terms
}
\description{
Constructs a P3 class
}
\details{
This function will take two integer values, x and k, and return the exponential of x to a sum of k terms
}
\examples{

}
